Creates reusable functions to style callouts
